Move UDQ action enum to UDQEnums
