+++ src/opm/parser/eclipse/EclipseState/Schedule/UDQ/UDQEnums.cpp
@@ -54,5 +54,22 @@ UDQVarType varType(const std::string& keyword) {
 
 }
 
+
+UDQAction actionType(const std::string& action_string) {
+    if (action_string == "ASSIGN")
+        return UDQAction::ASSIGN;
+
+    if (action_string == "DEFINE")
+        return UDQAction::DEFINE;
+
+    if (action_string == "UNITS")
+        return UDQAction::UNITS;
+
+    if (action_string == "UPDATE")
+        return UDQAction::UPDATE;
+
+    throw std::invalid_argument("Invalid action string " + action_string);
+}
+
 }
 }
